feat: update

+++ public/delete_product_group.sql
@@ -1,12 +1,16 @@
-CREATE PROCEDURE delete_product_group(IN param_group_id bigint)
+CREATE PROCEDURE delete_product_group(IN param_group_id bigint, INOUT result character varying DEFAULT NULL::character varying)
     LANGUAGE plpgsql
 AS
 $$
     BEGIN
-        DELETE FROM product_group where product_group_id = param_group_id;
-        RAISE NOTICE 'Successfully Deleted!';
+        IF NOT exists(SELECT*FROM product_group WHERE product_group_id = param_group_id) THEN
+                result:='Group not found!';
+        ELSE
+            DELETE FROM product_group WHERE product_group_id = param_group_id;
+            result:='Successfully Deleted!';
+       END IF;
     END
 $$;
 
-ALTER PROCEDURE delete_product_group(BIGINT) OWNER TO postgres;
+ALTER PROCEDURE delete_product_group(BIGINT, INOUT VARCHAR) OWNER TO postgres;
